training/requirements.txt
# ========================================
# TRAINING & DEVELOPMENT REQUIREMENTS
# Full environment for model training, data preprocessing, and local testing
# ========================================
#
# Installation:
#   pip install -r requirements-train.txt
#
# This includes everything from requirements.txt plus training tools
# ========================================

# API Server
fastapi==0.121.2
uvicorn==0.38.0
starlette==0.49.3
pydantic==2.12.4
pydantic_core==2.41.5
anyio==4.11.0
h11==0.16.0
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0

# Chess Engine
python-chess==1.999

# Neural Network (GPU-enabled for training)
torch>=2.0.0
numpy>=1.24.0
huggingface-hub>=0.20.0

# Data Download & Preprocessing
zstandard>=0.21.0  # PGN decompression from Lichess database
requests>=2.31.0   # HTTP downloads
tqdm>=4.65.0       # Progress bars during training
pyyaml>=6.0.0      # Training config files

# Modal Cloud Training (optional)
modal>=0.63.0

# Development Tools (optional but recommended)
ipython>=8.12.0
jupyter>=1.0.0
matplotlib>=3.7.0  # For training visualizations
